add separate commands for loading positions from pgn and wthor
